Car & Automobile Scraping/Taxi Dispatch System Tbmslive Drivers Information/requirements.txt
google_api_python_client==2.90.0
itemadapter==0.7.0
openpyxl==3.1.1
protobuf==4.23.3
Scrapy==2.8.0
scrapy_xlsx==0.1.1
